+++ src/components/modals/ReferenceModal.vue
@@ -7,6 +7,27 @@
       <font-awesome-icon icon="address-card" class="fa fa-address-card" />
       {{ edition.name || "Custom Script" }}
     </h3>
+    <h5 v-if="edition.version">{{ edition.version }}</h5>
+    <br>
+
+
+    <div v-if="edition.bootlegger" class="specialRuleContainer">
+      <img src="../../assets/icons/bootlegger.png" class="bootlegger left">
+      <img src="../../assets/icons/bootlegger.png" class="bootlegger right">
+      <div v-if="edition.bootlegger.length && edition.bootlegger.length > 0" class="specialRule">
+        <div v-for="elem in edition.bootlegger" :key="elem">
+          <div v-if="typeof elem == 'number' && elem >= 0 && elem < t('modal.reference.specialRules').length">
+            <p>{{ t('modal.reference.specialRules')[elem] }}</p>
+          </div>
+          <div v-else>
+            <p>{{ elem }}</p>
+          </div>
+        </div>
+      </div>
+      <div v-else class="specialRule">{{ edition.bootlegger }}</div>
+    </div>
+
+
     <div v-for="(teamRoles, team) in rolesGrouped" :key="team" :class="['team', team]">
       <aside :aria-label="t(`modal.reference.teamNames.${team}`)">
         <h4>{{ t(`modal.reference.teamNames.${team}`) }}</h4>
@@ -18,7 +39,10 @@
             <span v-if="Object.keys(playersByRole).length" class="player">
               {{ playersByRole[role.id] ? playersByRole[role.id]?.join(", ") : "" }}
             </span>
-            <span class="name">{{ role.name }}</span>
+            <div class="name">
+              {{ role.name }}
+              <img v-if="edition.stormcaught == role.id" src="../../assets/icons/stormcatcher.png" class="stormcatcher">
+            </div>
             <span class="ability">{{ role.ability }}</span>
           </div>
         </li>
@@ -143,6 +167,11 @@ h3 {
   }
 }
 
+h5 {
+  font-size: 70%;
+  color: gray;
+}
+
 .townsfolk {
   .name {
     color: $townsfolk;
@@ -182,6 +211,42 @@ h3 {
   width: 4rem;
   height: 4rem;
   text-align: right;
+}
+
+.stormcatcher {
+  width: 1.5rem;
+  position: absolute;
+  float: right;
+  margin-left: 15px;
+}
+
+div.container {
+  text-align: center;
+}
+
+
+.specialRuleContainer {
+  width: 100%;
+}
+
+.bootlegger {
+  width: 10rem;
+  display: inline-block;
+
+
+  &.left {
+    float: left;
+  }
+  &.right {
+    float: right;
+  }
+}
+
+.specialRule {
+  text-align: center;
+  color: $loric;
+  font-weight: bold;
+  font-size: 80%;
 }
 
 .outsider {
@@ -255,6 +320,7 @@ h3 {
 .team {
   display: flex;
   align-items: stretch;
+  clear: both;
 
   &:not(:last-child):after {
     content: " ";